main: extract Intune Chat domain into intune-chat/service.ts
Stage B of the state.ts decomposition: move chat send/stream, graph
cache, multi-tenant chat/batch, tenant groups, and self-training
methods into IntuneChatService behind an explicit ChatServiceHost
interface. AppStateStore keeps one-line delegates so the IPC surface
and tests are untouched. Verbatim moves; state.ts 6,916 -> 5,289.
