+++ apps/desktop/electron/state.ts
@@ -55,26 +55,19 @@ import type {
   TenantRecord,
   TenantSession,
   GraphCacheRefreshResult,
-  GraphCacheRefreshResourceResult,
   GraphCacheRefreshScheduleSettings,
-  GraphCacheResourceKind,
   GraphCacheStatus,
-  HostedProviderBatchConsent,
   ImportMultiTenantResultToWorkspacesInput,
   ImportMultiTenantResultToWorkspacesResult,
   CreateWorkspaceInput,
   IntuneChatConversation,
   IntuneChatMessage,
-  IntuneChatProgressStep,
-  IntuneChatStreamStage,
   IntuneChatStreamEvent,
   LocalDataSummary,
   MultiTenantAgentBatch,
   MultiTenantChatJob,
   MultiTenantChatRunResult,
   MultiTenantChatStreamEvent,
-  MultiTenantDeviceRow,
-  MultiTenantTenantComparison,
   PinWorkspaceEvidenceInput,
   PreflightMultiTenantChatInput,
   QueueMultiTenantAgentBatchInput,
@@ -97,8 +90,6 @@ import type {
   WorkspaceEvidence,
   WorkspaceLink,
   WorkspaceNote,
-  WorkspacePromptContextInput,
-  WorkspacePromptContextSummary,
   WorkspaceSummary,
 } from "@openadminos/agent-sdk";
 import {
@@ -135,28 +126,13 @@ import { SafeStorageConnectorSecretStore } from "./connector-secret-store.js";
 import { requestConnectorConfirmation } from "./connector-confirm-bridge.js";
 import { searchEndpoints } from "./graph-catalog.js";
 
-type WorkspacePromptContextPayload = {
-  summary: WorkspacePromptContextSummary;
-  promptBlock: string;
-};
-
 import {
   DEFAULT_REGISTRY_SOURCE,
   refreshRegistry,
   validateRegistrySource,
 } from "./registry-client.js";
 import { IntelligenceSqliteStore } from "./intune-chat/sqlite-store.js";
-import {
-  GRAPH_CACHE_RESOURCES,
-  buildAnswerPack,
-  chatTitleForPrompt,
-  definitionForResource,
-  matchAgentsToQuestion,
-  pathForResource,
-  planChatContext,
-  requiredScopesForResources,
-  selfTrainingCandidateFromPrompt,
-} from "./intune-chat/planner.js";
+import { IntuneChatService } from "./intune-chat/service.js";
 import {
   OUTLOOK_CONNECTOR_ID,
   SLACK_CONNECTOR_ID,
@@ -226,46 +202,15 @@ import {
 } from "./provider-detail.js";
 import {
   DEFAULT_STATS_API_URL,
-  assertConversationTenant,
-  buildChatProgressSteps,
-  buildIntuneChatSources,
-  buildIntuneChatSystemPrompt,
-  buildMultiTenantDossierMarkdown,
-  buildMultiTenantSummary,
-  buildSelfTrainingYaml,
-  buildTenantComparison,
   createLocalOnlyTenantSession,
-  emptyMultiTenantSummary,
   entryToRegistrySummary,
-  estimateChatProgressPercent,
-  fetchGraphCachePages,
-  hashTenantId,
   humanizeMsalError,
   humanizeScheduledRunError,
-  intuneChatProviderBudget,
-  isGraphCacheStatusStale,
   isNodeError,
   isTerminalRunStatus,
-  isWindowsCompliancePrompt,
-  newestRefreshedAt,
-  normalizeChatConversationTitle,
-  normalizeMultiTenantDeviceRow,
-  normalizeTenantGroupName,
   normalizeWorkspaceInstructions,
   normalizeWorkspaceTitle,
-  readPlannedChatRows,
-  readinessRecovery,
-  readinessWarnings,
-  resolveTenantGroups,
-  resolveTenantScopeIds,
-  runWithConcurrency,
-  sanitizeGraphResources,
-  stableSuggestionId,
-  summarizeMultiTenantResult,
   tenantNamesById,
-  tenantReadinessStatus,
-  trimForPrompt,
-  updateJobTenantProgress,
   withSelfTrainingOverlay,
 } from "./state-helpers.js";
 export { __agentDraftTestUtils } from "./agent-draft-helpers.js";
@@ -328,21 +273,6 @@ interface PersistedState {
   /** User-overridable registry source URL. */
   registrySource?: string;
 }
-
-type GraphCacheRefreshProgressEvent =
-  | {
-      type: "resource-start";
-      resource: GraphCacheResourceKind;
-      label: string;
-      completed: number;
-      total: number;
-    }
-  | {
-      type: "resource-complete";
-      result: GraphCacheRefreshResourceResult;
-      completed: number;
-      total: number;
-    };
 
 type WhatsAppWebClientLike = Pick<
   WhatsAppWebClient,
@@ -451,6 +381,7 @@ export class AppStateStore {
   private readonly sandboxedCodeDefault: boolean;
   private readonly userDataPath: string | undefined;
   private readonly intelligenceStore: IntelligenceSqliteStore | undefined;
+  private readonly chatService: IntuneChatService;
   private readonly graphFactory: AppStateStoreOptions["graphFactory"] | undefined;
   private readonly llmFactory: AppStateStoreOptions["llmFactory"] | undefined;
   private readonly whatsAppWebClientFactory:
@@ -524,6 +455,28 @@ export class AppStateStore {
         : undefined;
       this.allowDevRegistrySource = options.allowDevRegistrySource === true;
     }
+    const host = this;
+    this.chatService = new IntuneChatService({
+      read: () => host.read(),
+      requireIntelligenceStore: () => host.requireIntelligenceStore(),
+      resolveTenant: (persisted, tenantId) =>
+        host.resolveTenant(persisted as PersistedState, tenantId),
+      listProviders: () => host.listProviders(),
+      buildLlm: (providerId, model) => host.buildLlm(providerId, model),
+      buildGraph: (pinnedTenantId, agentScopes) =>
+        host.buildGraph(pinnedTenantId, agentScopes),
+      startRun: (agentSlug, options) => host.startRun(agentSlug, options),
+      appVersion: this.appVersion,
+      get userDataPath() {
+        return host.userDataPath;
+      },
+      get intelligenceStore() {
+        return host.intelligenceStore;
+      },
+      get graphFactory() {
+        return host.graphFactory;
+      },
+    });
     // Point the runtime at the OTA-updated manifest tree (if userData is
     // configured). When an agent has been updated via `updateAgent`, the
     // runtime resolves its manifest from here instead of the bundled tree.
@@ -728,80 +681,8 @@ export class AppStateStore {
     return tenant;
   }
 
-  private async maybeCreateSelfTrainingSuggestionFromChat(input: {
-    tenantId: string;
-    question: string;
-    agentSuggestions: NonNullable<IntuneChatMessage["agentSuggestions"]>;
-  }): Promise<void> {
-    const store = this.requireIntelligenceStore();
-    const settings = store.getSelfTrainingSettings();
-    if (!settings.enabled) return;
-    const candidate = selfTrainingCandidateFromPrompt(input);
-    if (!candidate) return;
-    const now = new Date().toISOString();
-    const id = stableSuggestionId(input.tenantId, candidate.agentSlug, candidate.text);
-    store.recordLearningEvent({
-      id: `event_${randomUUID()}`,
-      tenantId: input.tenantId,
-      agentSlug: candidate.agentSlug,
-      eventType: "chat.preference-detected",
-      source: "chat",
-      payload: { question: input.question, candidate },
-      createdAt: now,
-    });
-    store.createSelfTrainingSuggestion({
-      id,
-      tenantId: input.tenantId,
-      agentSlug: candidate.agentSlug,
-      status: "pending",
-      text: candidate.text,
-      reason: candidate.reason,
-      source: "chat",
-      createdAt: now,
-    });
-  }
-
-  private async writeSelfTrainingFile(
-    tenantId: string,
-    agentSlug: string,
-  ): Promise<void> {
-    if (!this.userDataPath) return;
-    const store = this.requireIntelligenceStore();
-    const accepted = store.listAcceptedSelfTrainingSuggestions({ tenantId, agentSlug });
-    const tenantKey = hashTenantId(tenantId);
-    const dir = join(
-      this.userDataPath,
-      "agent-learning",
-      "tenants",
-      tenantKey,
-      "agents",
-      agentSlug,
-    );
-    await mkdir(dir, { recursive: true });
-    const updatedAt = new Date().toISOString();
-    const yaml = buildSelfTrainingYaml({
-      agentSlug,
-      tenantKey,
-      updatedAt,
-      suggestions: accepted,
-    });
-    await writeFile(join(dir, "self-training.yaml"), yaml, "utf8");
-  }
-
   private selfTrainingPromptOverlay(tenantId: string, agentSlug: string): string | undefined {
-    if (!this.intelligenceStore) return undefined;
-    const settings = this.intelligenceStore.getSelfTrainingSettings();
-    if (!settings.enabled) return undefined;
-    const accepted = this.intelligenceStore.listAcceptedSelfTrainingSuggestions({
-      tenantId,
-      agentSlug,
-    });
-    if (accepted.length === 0) return undefined;
-    return [
-      "Approved local self-training instructions for this tenant and agent:",
-      ...accepted.map((entry, index) => `${index + 1}. ${entry.text}`),
-      "These instructions may guide reasoning and wording only. They do not change scopes, write mode, connector egress, or confirmation policy.",
-    ].join("\n");
+    return this.chatService.selfTrainingPromptOverlay(tenantId, agentSlug);
   }
 
   private recordLearningEventSafely(input: {
@@ -811,20 +692,7 @@ export class AppStateStore {
     source: string;
     payload: unknown;
   }): void {
-    if (!this.intelligenceStore || !input.tenantId) return;
-    try {
-      this.intelligenceStore.recordLearningEvent({
-        id: `event_${randomUUID()}`,
-        tenantId: input.tenantId,
-        agentSlug: input.agentSlug,
-        eventType: input.eventType,
-        source: input.source,
-        payload: input.payload,
-        createdAt: new Date().toISOString(),
-      });
-    } catch (error) {
-      console.warn("[self-training] failed to record learning event", error);
-    }
+    this.chatService.recordLearningEventSafely(input);
   }
 
   private serialize<T>(task: () => Promise<T>): Promise<T> {
@@ -1011,58 +879,37 @@ export class AppStateStore {
   }
 
   async listIntuneChatConversations(): Promise<IntuneChatConversation[]> {
-    return this.requireIntelligenceStore().listConversations();
+    return this.chatService.listIntuneChatConversations();
   }
 
   async searchIntuneChatConversations(query: string): Promise<IntuneChatConversation[]> {
-    return this.requireIntelligenceStore().searchConversations(query);
+    return this.chatService.searchIntuneChatConversations(query);
   }
 
   async renameIntuneChatConversation(
     conversationId: string,
     title: string,
   ): Promise<IntuneChatConversation> {
-    if (!conversationId.trim()) {
-      throw new Error("Conversation id is required.");
-    }
-    const normalizedTitle = normalizeChatConversationTitle(title);
-    return this.requireIntelligenceStore().renameConversation(
-      conversationId,
-      normalizedTitle,
-      new Date().toISOString(),
-    );
+    return this.chatService.renameIntuneChatConversation(conversationId, title);
   }
 
   async setIntuneChatConversationPinned(
     conversationId: string,
     pinned: boolean,
   ): Promise<IntuneChatConversation> {
-    if (!conversationId.trim()) {
-      throw new Error("Conversation id is required.");
-    }
-    return this.requireIntelligenceStore().setConversationPinned(
-      conversationId,
-      pinned,
-      new Date().toISOString(),
-    );
+    return this.chatService.setIntuneChatConversationPinned(conversationId, pinned);
   }
 
   async deleteIntuneChatConversation(conversationId: string): Promise<void> {
-    if (!conversationId.trim()) {
-      throw new Error("Conversation id is required.");
-    }
-    this.requireIntelligenceStore().deleteConversation(conversationId);
+    return this.chatService.deleteIntuneChatConversation(conversationId);
   }
 
   async getIntuneChatMessages(conversationId: string): Promise<IntuneChatMessage[]> {
-    if (!conversationId.trim()) {
-      throw new Error("Conversation id is required.");
-    }
-    return this.requireIntelligenceStore().listMessages(conversationId);
+    return this.chatService.getIntuneChatMessages(conversationId);
   }
 
   async listTenantGroups(): Promise<TenantGroup[]> {
-    return this.requireIntelligenceStore().listTenantGroups();
+    return this.chatService.listTenantGroups();
   }
 
   async saveTenantGroup(input: {
@@ -1070,36 +917,21 @@ export class AppStateStore {
     name: string;
     tenantIds: string[];
   }): Promise<TenantGroup> {
-    const persisted = await this.read();
-    const knownTenantIds = new Set(persisted.tenants.map((tenant) => tenant.id));
-    const tenantIds = [...new Set(input.tenantIds)].filter((tenantId) => {
-      if (!knownTenantIds.has(tenantId)) {
-        throw new Error(`Tenant group includes an unknown tenant: ${tenantId}`);
-      }
-      return true;
-    });
-    const name = normalizeTenantGroupName(input.name);
-    return this.requireIntelligenceStore().saveTenantGroup({
-      id: input.id?.trim() || `tgrp_${randomUUID()}`,
-      name,
-      tenantIds,
-      now: new Date().toISOString(),
-    });
+    return this.chatService.saveTenantGroup(input);
   }
 
   async deleteTenantGroup(id: string): Promise<void> {
-    if (!id.trim()) throw new Error("Tenant group id is required.");
-    this.requireIntelligenceStore().deleteTenantGroup(id);
+    return this.chatService.deleteTenantGroup(id);
   }
 
   async listSavedMultiTenantQueries(): Promise<SavedMultiTenantQuery[]> {
-    return this.requireIntelligenceStore().listSavedMultiTenantQueries();
+    return this.chatService.listSavedMultiTenantQueries();
   }
 
   async preflightMultiTenantIntuneChat(
     input: PreflightMultiTenantChatInput,
   ): Promise<TenantScopePreflight> {
-    return this.buildMultiTenantPreflight(input);
+    return this.chatService.preflightMultiTenantIntuneChat(input);
   }
 
   async runMultiTenantIntuneChat(
@@ -1107,288 +939,7 @@ export class AppStateStore {
     onEvent?: (event: MultiTenantChatStreamEvent) => void,
     options: { signal?: AbortSignal } = {},
   ): Promise<MultiTenantChatRunResult> {
-    const content = input.prompt.trim();
-    if (!content) throw new Error("Multi-tenant chat prompt is required.");
-    const store = this.requireIntelligenceStore();
-    const persisted = await this.read();
-    const preflight = await this.buildMultiTenantPreflight(input);
-    const providers = await this.listProviders();
-    const activeTenant = this.resolveTenant(persisted);
-    const provider =
-      providers.find((entry) => entry.id === persisted.activeProviderId) ?? providers[0];
-    const providerId = provider?.id ?? persisted.activeProviderId;
-    this.requireHostedBatchConsent(input.hostedProviderConsent, preflight, provider, providerId);
-
-    const now = new Date().toISOString();
-    const jobId = `mtjob_${randomUUID()}`;
-    let job: MultiTenantChatJob = {
-      id: jobId,
-      prompt: content,
-      ...(input.savedQueryId ? { savedQueryId: input.savedQueryId } : {}),
-      tenantScope: preflight.tenantScope,
-      resolvedTenantIds: preflight.resolvedTenantIds,
-      providerId,
-      providerName: preflight.providerName,
-      providerIsLocal: preflight.providerIsLocal,
-      ...(preflight.model ? { model: preflight.model } : {}),
-      status: "running",
-      createdAt: now,
-      updatedAt: now,
-      preflight,
-      progress: preflight.tenants.map((tenant) => ({
-        tenantId: tenant.tenantId,
-        tenantName: tenant.tenantName,
-        status:
-          tenant.status === "expired" ||
-          tenant.status === "missing-scopes" ||
-          tenant.status === "throttled" ||
-          tenant.status === "failed"
-            ? "skipped"
-            : "queued",
-        detail: tenant.recovery,
-        updatedAt: now,
-      })),
-      summary: emptyMultiTenantSummary(),
-      comparisons: [],
-      deviceRows: [],
-      assistantText: "",
-      exportDossierMarkdown: "",
-    };
-    const persistJob = (eventType: "started" | "progress" = "progress") => {
-      store.upsertMultiTenantJob(job);
-      onEvent?.({ type: eventType, job });
-    };
-    const assertNotCancelled = () => {
-      if (options.signal?.aborted !== true) return;
-      job = {
-        ...job,
-        status: "cancelled",
-        error: "Stopped by user.",
-        updatedAt: new Date().toISOString(),
-      };
-      store.upsertMultiTenantJob(job);
-      onEvent?.({ type: "cancelled", job });
-      throw new Error("Multi-tenant chat run stopped.");
-    };
-    persistJob("started");
-
-    const runnable = preflight.tenants.filter(
-      (tenant) =>
-        tenant.selected &&
-        tenant.status !== "expired" &&
-        tenant.status !== "missing-scopes" &&
-        tenant.status !== "throttled" &&
-        tenant.status !== "failed" &&
-        tenant.status !== "skipped",
-    );
-    const refreshTenant = async (tenant: (typeof runnable)[number]) => {
-      const startedAt = new Date().toISOString();
-      job = updateJobTenantProgress(job, tenant.tenantId, {
-        status: input.refreshIfStale === false ? "reading-cache" : "refreshing-cache",
-        detail:
-          input.refreshIfStale === false
-            ? "Reading existing local cache."
-            : "Refreshing prompt-relevant cache.",
-        updatedAt: startedAt,
-      });
-      persistJob();
-      try {
-        assertNotCancelled();
-        if (input.refreshIfStale !== false && tenant.staleResources.length > 0) {
-          const refreshResult = await this.refreshGraphCacheInternal({
-            tenantId: tenant.tenantId,
-            resources: tenant.staleResources,
-          });
-          const failedResources = refreshResult.resources.filter((resource) => !resource.ok);
-          if (
-            failedResources.some((resource) => resource.resource === "managedDevices") ||
-            failedResources.length === refreshResult.resources.length
-          ) {
-            throw new Error(
-              failedResources
-                .map(
-                  (resource) =>
-                    `${resource.label}: ${resource.error ?? "Graph cache refresh failed."}`,
-                )
-                .join("; "),
-            );
-          }
-        }
-        job = updateJobTenantProgress(job, tenant.tenantId, {
-          status: "building-result",
-          detail: "Computing tenant result from local cache.",
-          updatedAt: new Date().toISOString(),
-        });
-        persistJob();
-      } catch (error) {
-        if (options.signal?.aborted === true) {
-          throw error;
-        }
-        const message = error instanceof Error ? error.message : String(error);
-        job = updateJobTenantProgress(job, tenant.tenantId, {
-          status: "failed",
-          detail: message,
-          updatedAt: new Date().toISOString(),
-        });
-        persistJob();
-      }
-    };
-    assertNotCancelled();
-    await runWithConcurrency(runnable, 2, refreshTenant);
-    assertNotCancelled();
-
-    const comparisons: MultiTenantTenantComparison[] = [];
-    const deviceRows: MultiTenantDeviceRow[] = [];
-    for (const tenant of preflight.tenants) {
-      if (!tenant.selected) continue;
-      const failedProgress = job.progress.find(
-        (entry) => entry.tenantId === tenant.tenantId && entry.status === "failed",
-      );
-      if (failedProgress) {
-        comparisons.push({
-          tenantId: tenant.tenantId,
-          tenantName: tenant.tenantName,
-          status: "failed",
-          windowsDevices: 0,
-          compliant: 0,
-          nonCompliant: 0,
-          unknown: 0,
-          lastRefresh: tenant.cacheFreshness,
-          warnings: [...tenant.warnings, failedProgress.detail ?? "Tenant refresh failed."],
-        });
-        continue;
-      }
-      if (!runnable.some((entry) => entry.tenantId === tenant.tenantId)) {
-        comparisons.push({
-          tenantId: tenant.tenantId,
-          tenantName: tenant.tenantName,
-          status: tenant.status,
-          windowsDevices: 0,
-          compliant: 0,
-          nonCompliant: 0,
-          unknown: 0,
-          lastRefresh: tenant.cacheFreshness,
-          warnings: tenant.warnings,
-        });
-        continue;
-      }
-      const tenantRows = store.readManagedDeviceRowsForTenant({
-        tenantId: tenant.tenantId,
-        limit: 10_000,
-      });
-      const normalizedRows = tenantRows
-        .map((entry) =>
-          normalizeMultiTenantDeviceRow({
-            row: entry.row,
-            tenantId: tenant.tenantId,
-            tenantName: tenant.tenantName,
-            sourceRefreshedAt: entry.refreshedAt,
-          }),
-        )
-        .filter((row): row is MultiTenantDeviceRow => Boolean(row));
-      const windowsRows = normalizedRows.filter((row) =>
-        row.operatingSystem.toLowerCase().includes("windows"),
-      );
-      deviceRows.push(...windowsRows);
-      comparisons.push(buildTenantComparison({
-        tenant,
-        rows: windowsRows,
-      }));
-      job = updateJobTenantProgress(job, tenant.tenantId, {
-        status: "ready",
-        detail: `${windowsRows.length.toLocaleString()} Windows device rows prepared.`,
-        updatedAt: new Date().toISOString(),
-      });
-      persistJob();
-    }
-
-    const summary = buildMultiTenantSummary(comparisons);
-    const assistantText = await this.buildMultiTenantAssistantText({
-      prompt: content,
-      providerId,
-      provider,
-      model: preflight.model,
-      summary,
-      comparisons,
-    });
-    const exportDossierMarkdown = buildMultiTenantDossierMarkdown({
-      prompt: content,
-      providerName: preflight.providerName,
-      model: preflight.model,
-      generatedAt: new Date().toISOString(),
-      summary,
-      comparisons,
-      rows: deviceRows,
-    });
-    const finalStatus =
-      comparisons.some((comparison) =>
-        ["failed", "expired", "missing-scopes", "throttled", "skipped"].includes(
-          comparison.status,
-        ),
-      )
-        ? "partial"
-        : "completed";
-    job = {
-      ...job,
-      status: finalStatus,
-      summary,
-      comparisons,
-      deviceRows,
-      assistantText,
-      exportDossierMarkdown,
-      updatedAt: new Date().toISOString(),
-    };
-
-    const conversation = store.createConversation({
-      id: `chat_${randomUUID()}`,
-      title: chatTitleForPrompt(content),
-      tenantId: activeTenant.id,
-      now,
-      scopeKind: "multi-tenant",
-      tenantScope: preflight.tenantScope,
-      multiTenantJobId: job.id,
-    });
-    job = { ...job, conversationId: conversation.id, updatedAt: new Date().toISOString() };
-    persistJob();
-    const userMessage: IntuneChatMessage = {
-      id: `msg_${randomUUID()}`,
-      conversationId: conversation.id,
-      role: "user",
-      content,
-      status: "completed",
-      createdAt: now,
-    };
-    const assistantMessage: IntuneChatMessage = {
-      id: `msg_${randomUUID()}`,
-      conversationId: conversation.id,
-      role: "assistant",
-      content: assistantText,
-      status: finalStatus === "completed" || finalStatus === "partial" ? "completed" : "failed",
-      createdAt: new Date().toISOString(),
-      providerId,
-      ...(preflight.model ? { model: preflight.model } : {}),
-      sources: [
-        {
-          resource: "managedDevices",
-          label: "Intune managed devices",
-          rows: deviceRows.length,
-          source: "cache",
-          path: "/deviceManagement/managedDevices",
-        },
-      ],
-    };
-    store.insertMessage(userMessage);
-    store.insertMessage(assistantMessage);
-    store.touchConversation(conversation.id, undefined, assistantMessage.createdAt);
-
-    const result = {
-      job,
-      conversation: store.getConversation(conversation.id) ?? conversation,
-      userMessage,
-      assistantMessage,
-    };
-    onEvent?.({ type: "completed", result });
-    return result;
+    return this.chatService.runMultiTenantIntuneChat(input, onEvent, options);
   }
 
   async streamMultiTenantIntuneChat(
@@ -1396,736 +947,71 @@ export class AppStateStore {
     onEvent: (event: MultiTenantChatStreamEvent) => void,
     options: { signal?: AbortSignal } = {},
   ): Promise<MultiTenantChatRunResult> {
-    try {
-      return await this.runMultiTenantIntuneChat(input, onEvent, options);
-    } catch (caught) {
-      if (options.signal?.aborted !== true) {
-        onEvent({
-          type: "failed",
-          error: caught instanceof Error ? caught.message : String(caught),
-        });
-      }
-      throw caught;
-    }
-  }
-
-  private buildWorkspacePromptContext(
-    input: WorkspacePromptContextInput,
-    tenantId: string,
-    persisted: PersistedState,
-  ): WorkspacePromptContextPayload {
-    const workspaceId = input.workspaceId.trim();
-    if (!workspaceId) throw new Error("Workspace context requires a workspace id.");
-    const store = this.requireIntelligenceStore();
-    const workspace = store.getWorkspace(workspaceId, tenantNamesById(persisted.tenants));
-    if (!workspace) throw new Error(`Workspace not found: ${workspaceId}`);
-    if (workspace.tenantId !== tenantId) {
-      throw new Error("Workspace context tenant does not match the active chat tenant.");
-    }
-
-    const selectedEvidenceIds = new Set(input.evidenceIds ?? []);
-    const selectedNoteIds = new Set(input.noteIds ?? []);
-    const evidence =
-      selectedEvidenceIds.size > 0
-        ? workspace.evidence.filter((entry) => selectedEvidenceIds.has(entry.id))
-        : [];
-    const notes =
-      selectedNoteIds.size > 0
-        ? workspace.notes.filter((entry) => selectedNoteIds.has(entry.id))
-        : [];
-    const includesInstructions =
-      input.includeInstructions === true && Boolean(workspace.instructions?.trim());
-
-    if (!includesInstructions && evidence.length === 0 && notes.length === 0) {
-      throw new Error("Select workspace evidence, notes, or instructions before attaching context.");
-    }
-
-    const missingEvidence = [...selectedEvidenceIds].filter(
-      (id) => !workspace.evidence.some((entry) => entry.id === id),
-    );
-    const missingNotes = [...selectedNoteIds].filter(
-      (id) => !workspace.notes.some((entry) => entry.id === id),
-    );
-    if (missingEvidence.length > 0 || missingNotes.length > 0) {
-      throw new Error("Workspace context includes evidence or notes that no longer exist.");
-    }
-
-    const lines = [
-      "Workspace context attached by the admin.",
-      `Workspace: ${workspace.title}`,
-      `Tenant: ${workspace.tenantName ?? workspace.tenantId}`,
-    ];
-    if (includesInstructions && workspace.instructions) {
-      lines.push("", "Workspace instructions:", workspace.instructions.trim());
-    }
-    if (notes.length > 0) {
-      lines.push("", "Workspace notes:");
-      notes.slice(0, 12).forEach((note, index) => {
-        lines.push(`${index + 1}. ${trimForPrompt(note.content, 1400)}`);
-      });
-    }
-    if (evidence.length > 0) {
-      lines.push("", "Workspace evidence:");
-      evidence.slice(0, 8).forEach((entry, index) => {
-        lines.push(
-          `${index + 1}. ${entry.title} (${entry.sourceType}, ${entry.createdAt})`,
-          trimForPrompt(JSON.stringify(entry.content, null, 2), 2800),
-        );
-      });
-    }
-
-    return {
-      summary: {
-        workspaceId: workspace.id,
-        workspaceTitle: workspace.title,
-        tenantId: workspace.tenantId,
-        evidenceCount: evidence.length,
-        noteCount: notes.length,
-        includesInstructions,
-      },
-      promptBlock: lines.join("\n"),
-    };
+    return this.chatService.streamMultiTenantIntuneChat(input, onEvent, options);
   }
 
   async listMultiTenantChatJobs(): Promise<MultiTenantChatJob[]> {
-    return this.requireIntelligenceStore().listMultiTenantJobs();
+    return this.chatService.listMultiTenantChatJobs();
   }
 
   async getMultiTenantChatJob(id: string): Promise<MultiTenantChatJob | undefined> {
-    if (!id.trim()) throw new Error("Multi-tenant job id is required.");
-    return this.requireIntelligenceStore().getMultiTenantJob(id);
+    return this.chatService.getMultiTenantChatJob(id);
   }
 
   async queueMultiTenantAgentBatch(
     input: QueueMultiTenantAgentBatchInput,
   ): Promise<QueueMultiTenantAgentBatchResult> {
-    const agentSlug = input.agentSlug.trim();
-    if (!agentSlug) throw new Error("Agent slug is required.");
-    const store = this.requireIntelligenceStore();
-    const persisted = await this.read();
-    const agent = persisted.installedAgents.find((entry) => entry.slug === agentSlug);
-    if (!agent) {
-      throw new Error(`Agent is not installed: ${agentSlug}`);
-    }
-    assertAgentCompatible(withAgentCompatibility(agent, this.appVersion), "run");
-
-    const prompt =
-      input.prompt?.trim() || `Run ${agent.name} against the selected tenant scope.`;
-    const preflight = await this.buildMultiTenantPreflight({
-      prompt,
-      tenantScope: input.tenantScope,
-      ...(input.savedQueryId ? { savedQueryId: input.savedQueryId } : {}),
-    });
-    const runnable = preflight.tenants.filter(
-      (tenant) =>
-        tenant.selected &&
-        tenant.status !== "expired" &&
-        tenant.status !== "missing-scopes" &&
-        tenant.status !== "throttled" &&
-        tenant.status !== "failed" &&
-        tenant.status !== "skipped",
-    );
-    if (runnable.length === 0) {
-      throw new Error("No selected tenants are ready for this agent batch.");
-    }
-
-    const now = new Date().toISOString();
-    let batch: MultiTenantAgentBatch = {
-      id: `mtbatch_${randomUUID()}`,
-      agentSlug: agent.slug,
-      agentName: agent.name,
-      agentMode: agent.mode,
-      tenantScope: preflight.tenantScope,
-      resolvedTenantIds: runnable.map((tenant) => tenant.tenantId),
-      status: "queued",
-      runIds: [],
-      createdAt: now,
-      updatedAt: now,
-      preflight,
-    };
-    store.upsertMultiTenantAgentBatch(batch);
-
-    const runs: RunRecord[] = [];
-    const failures: string[] = [];
-    for (const tenant of runnable) {
-      try {
-        const run = await this.startRun(agent.slug, {
-          tenantId: tenant.tenantId,
-          trigger: "manual",
-        });
-        runs.push(run);
-        batch = {
-          ...batch,
-          status: agent.mode === "write" ? "awaiting-confirmation" : "running",
-          runIds: runs.map((entry) => entry.id),
-          updatedAt: new Date().toISOString(),
-        };
-        store.upsertMultiTenantAgentBatch(batch);
-      } catch (caught) {
-        failures.push(
-          `${tenant.tenantName}: ${caught instanceof Error ? caught.message : String(caught)}`,
-        );
-      }
-    }
-
-    const finalStatus =
-      failures.length === 0
-        ? agent.mode === "write"
-          ? "awaiting-confirmation"
-          : "running"
-        : runs.length > 0
-          ? "partial"
-          : "failed";
-    batch = {
-      ...batch,
-      status: finalStatus,
-      runIds: runs.map((entry) => entry.id),
-      ...(failures.length > 0 ? { error: failures.join("; ") } : {}),
-      updatedAt: new Date().toISOString(),
-    };
-    store.upsertMultiTenantAgentBatch(batch);
-    return { batch, runs };
+    return this.chatService.queueMultiTenantAgentBatch(input);
   }
 
   async listMultiTenantAgentBatches(): Promise<MultiTenantAgentBatch[]> {
-    return this.requireIntelligenceStore().listMultiTenantAgentBatches();
+    return this.chatService.listMultiTenantAgentBatches();
   }
 
   async getMultiTenantAgentBatch(
     id: string,
   ): Promise<MultiTenantAgentBatch | undefined> {
-    if (!id.trim()) throw new Error("Multi-tenant agent batch id is required.");
-    return this.requireIntelligenceStore().getMultiTenantAgentBatch(id);
-  }
-
-  private async buildMultiTenantPreflight(
-    input: PreflightMultiTenantChatInput,
-  ): Promise<TenantScopePreflight> {
-    const prompt = input.prompt.trim();
-    if (!prompt) throw new Error("Multi-tenant chat prompt is required.");
-    const store = this.requireIntelligenceStore();
-    const persisted = await this.read();
-    const activeTenant = this.resolveTenant(persisted);
-    if (persisted.tenants.length === 0) {
-      throw new Error("Connect at least one tenant before using multi-tenant Intune Chat.");
-    }
-    const groups = store.listTenantGroups();
-    const resolvedGroups = resolveTenantGroups(input.tenantScope, groups);
-    const resolvedTenantIds = resolveTenantScopeIds({
-      scope: input.tenantScope,
-      groups: resolvedGroups,
-      tenants: persisted.tenants,
-      activeTenantId: activeTenant.id,
-    });
-    const selectedTenantSet = new Set(resolvedTenantIds);
-    const savedQuery = input.savedQueryId
-      ? store
-          .listSavedMultiTenantQueries()
-          .find((query) => query.id === input.savedQueryId)
-      : undefined;
-    const planned = planChatContext(prompt);
-    const resourceSet = new Set<GraphCacheResourceKind>([
-      ...planned.resources,
-      ...(savedQuery?.resourceHints ?? []),
-    ]);
-    if (isWindowsCompliancePrompt(prompt)) {
-      resourceSet.add("managedDevices");
-    }
-    const resources = sanitizeGraphResources([...resourceSet]);
-    const requiredScopes = requiredScopesForResources(resources);
-    const providers = await this.listProviders();
-    const provider =
-      providers.find((entry) => entry.id === persisted.activeProviderId) ?? providers[0];
-    const providerId = provider?.id ?? persisted.activeProviderId;
-    const model = resolveProviderDefaultModel(
-      provider,
-      persisted.activeModelByProviderId,
-    ).model;
-    const now = new Date().toISOString();
-    const tenants = persisted.tenants
-      .filter((tenant) => selectedTenantSet.has(tenant.id))
-      .map((tenant) => {
-        const statusRows = store.getGraphCacheStatus(tenant.id, [...GRAPH_CACHE_RESOURCES]);
-        const relevantStatus = statusRows.filter((status) =>
-          resources.includes(status.resource),
-        );
-        const staleResources = relevantStatus
-          .filter((status) => isGraphCacheStatusStale(status))
-          .map((status) => status.resource);
-        const cacheFreshness = newestRefreshedAt(relevantStatus);
-        const errors = relevantStatus
-          .map((status) => status.lastError)
-          .filter((error): error is string => Boolean(error));
-        const missingScopes = errors.some((error) => /scope|consent|permission/i.test(error))
-          ? requiredScopes
-          : [];
-        const status = tenantReadinessStatus({
-          missingScopes,
-          staleResources,
-          errors,
-          hasRows: relevantStatus.some((entry) => entry.rows > 0),
-        });
-        return {
-          tenantId: tenant.id,
-          tenantName: tenant.displayName,
-          username: tenant.username,
-          status,
-          selected: true,
-          cacheFreshness,
-          staleResources,
-          missingScopes,
-          warnings: readinessWarnings({ status, staleResources, errors }),
-          recovery: readinessRecovery(status),
-        };
-      });
-    return {
-      id: `preflight_${randomUUID()}`,
-      prompt,
-      tenantScope: input.tenantScope,
-      resolvedTenantIds,
-      resolvedGroups,
-      resources,
-      providerId,
-      providerName: provider?.name ?? providerId,
-      providerIsLocal: provider?.isLocal !== false,
-      ...(model ? { model } : {}),
-      generatedAt: now,
-      tenants,
-      canRun:
-        tenants.length > 0 &&
-        tenants.some((tenant) =>
-          tenant.status === "ready" || tenant.status === "stale",
-        ),
-    };
-  }
-
-  private requireHostedBatchConsent(
-    consent: HostedProviderBatchConsent | undefined,
-    preflight: TenantScopePreflight,
-    provider: ProviderSummary | undefined,
-    providerId: ProviderId,
-  ): void {
-    if (provider?.isLocal !== false) return;
-    if (!consent) {
-      throw new Error(
-        "Hosted provider confirmation is required before multi-tenant chat can send tenant context to the selected provider.",
-      );
-    }
-    if (consent.providerId !== providerId) {
-      throw new Error("Hosted provider confirmation does not match the selected provider.");
-    }
-    const expected = new Set(preflight.resolvedTenantIds);
-    const acknowledged = new Set(consent.tenantIds);
-    for (const tenantId of expected) {
-      if (!acknowledged.has(tenantId)) {
-        throw new Error("Hosted provider confirmation does not cover every selected tenant.");
-      }
-    }
-    const acknowledgedAt = Date.parse(consent.acknowledgedAt);
-    const now = Date.now();
-    if (
-      !Number.isFinite(acknowledgedAt) ||
-      now - acknowledgedAt > 5 * 60 * 1000 ||
-      acknowledgedAt - now > 60 * 1000
-    ) {
-      throw new Error("Hosted provider confirmation expired. Confirm the batch again.");
-    }
-  }
-
-  private async buildMultiTenantAssistantText(input: {
-    prompt: string;
-    providerId: ProviderId;
-    provider: ProviderSummary | undefined;
-    model?: string;
-    summary: MultiTenantChatJob["summary"];
-    comparisons: MultiTenantChatJob["comparisons"];
-  }): Promise<string> {
-    const deterministic = summarizeMultiTenantResult(input.summary, input.comparisons);
-    const llm = await this.buildLlm(input.providerId, input.model);
-    if (!llm.available) return deterministic;
-    try {
-      const completion = await llm.complete({
-        system: buildIntuneChatSystemPrompt(input.provider?.isLocal === true),
-        prompt: [
-          "Summarize this read-only multi-tenant Intune Chat result for an admin.",
-          "The JSON table is the source of truth. Mention partial, failed, skipped, or stale tenants.",
-          `Question: ${input.prompt}`,
-          JSON.stringify(
-            {
-              summary: input.summary,
-              tenants: input.comparisons,
-            },
-            null,
-            2,
-          ),
-        ].join("\n\n"),
-        ...(input.model ? { model: input.model } : {}),
-        temperature: 0.2,
-        maxTokens: 700,
-      });
-      return completion.text.trim() || deterministic;
-    } catch (error) {
-      const message = error instanceof Error ? error.message : String(error);
-      return `${deterministic}\n\nThe selected LLM provider failed while summarizing this multi-tenant result. ${message}`;
-    }
+    return this.chatService.getMultiTenantAgentBatch(id);
   }
 
   async getGraphCacheStatus(tenantId?: string): Promise<GraphCacheStatus> {
-    const persisted = await this.read();
-    const resolvedTenant = this.resolveTenant(persisted, tenantId);
-    const store = this.requireIntelligenceStore();
-    return {
-      tenantId: resolvedTenant.id,
-      resources: store.getGraphCacheStatus(
-        resolvedTenant.id,
-        [...GRAPH_CACHE_RESOURCES],
-      ),
-      schedule: store.getGraphCacheRefreshSchedule(resolvedTenant.id),
-    };
+    return this.chatService.getGraphCacheStatus(tenantId);
   }
 
   async getGraphCacheRefreshSchedule(
     tenantId?: string,
   ): Promise<GraphCacheRefreshScheduleSettings> {
-    const persisted = await this.read();
-    const resolvedTenant = this.resolveTenant(persisted, tenantId);
-    return this.requireIntelligenceStore().getGraphCacheRefreshSchedule(
-      resolvedTenant.id,
-    );
+    return this.chatService.getGraphCacheRefreshSchedule(tenantId);
   }
 
   async setGraphCacheRefreshSchedule(
     input: SetGraphCacheRefreshScheduleInput,
   ): Promise<GraphCacheRefreshScheduleSettings> {
-    const persisted = await this.read();
-    const resolvedTenant = this.resolveTenant(persisted, input.tenantId);
-    const intervalMinutes =
-      typeof input.intervalMinutes === "number" && Number.isFinite(input.intervalMinutes)
-        ? input.intervalMinutes
-        : 360;
-    return this.requireIntelligenceStore().setGraphCacheRefreshSchedule({
-      tenantId: resolvedTenant.id,
-      enabled: input.enabled,
-      intervalMinutes,
-      now: new Date().toISOString(),
-    });
+    return this.chatService.setGraphCacheRefreshSchedule(input);
   }
 
   async getLocalDataSummary(tenantId?: string): Promise<LocalDataSummary> {
-    const persisted = await this.read();
-    const resolvedTenant =
-      tenantId || persisted.activeTenantId
-        ? this.resolveTenant(persisted, tenantId)
-        : undefined;
-    return this.requireIntelligenceStore().getLocalDataSummary({
-      tenantId: resolvedTenant?.id,
-      definitions: resolvedTenant ? [...GRAPH_CACHE_RESOURCES] : undefined,
-    });
+    return this.chatService.getLocalDataSummary(tenantId);
   }
 
   async clearIntuneChatHistory(): Promise<LocalDataSummary> {
-    this.requireIntelligenceStore().clearChatHistory();
-    return this.getLocalDataSummary();
+    return this.chatService.clearIntuneChatHistory();
   }
 
   async clearGraphCache(tenantId?: string): Promise<LocalDataSummary> {
-    const persisted = await this.read();
-    const resolvedTenant = this.resolveTenant(persisted, tenantId);
-    this.requireIntelligenceStore().clearGraphCache(resolvedTenant.id);
-    return this.getLocalDataSummary(resolvedTenant.id);
+    return this.chatService.clearGraphCache(tenantId);
   }
 
   async refreshGraphCache(
     options: RefreshGraphCacheOptions = {},
   ): Promise<GraphCacheRefreshResult> {
-    return this.refreshGraphCacheInternal(options);
-  }
-
-  private async refreshGraphCacheInternal(
-    options: RefreshGraphCacheOptions = {},
-    onProgress?: (event: GraphCacheRefreshProgressEvent) => void,
-  ): Promise<GraphCacheRefreshResult> {
-    const store = this.requireIntelligenceStore();
-    const persisted = await this.read();
-    const tenant = this.resolveTenant(persisted, options.tenantId);
-    const resources = sanitizeGraphResources(options.resources);
-    const scopes = requiredScopesForResources(resources);
-    const startedAt = new Date().toISOString();
-    const log = (level: RunLogLevel, message: string, metadata?: Record<string, unknown>) => {
-      console.info("[intune-chat][graph]", level, message, metadata ?? "");
-    };
-    const graph = this.graphFactory
-      ? this.graphFactory({ tenantId: tenant.id, scopes, log })
-      : (await this.buildGraph(tenant.id, scopes)).createGraph(log);
-    const results: GraphCacheRefreshResult["resources"] = [];
-
-    for (const resource of resources) {
-      const definition = definitionForResource(resource);
-      const request = pathForResource(resource);
-      const refreshedAt = new Date().toISOString();
-      onProgress?.({
-        type: "resource-start",
-        resource,
-        label: definition.label,
-        completed: results.length,
-        total: resources.length,
-      });
-      try {
-        const query: Record<string, string> = { ...(request.query ?? {}) };
-        if (request.select && request.select.length > 0) {
-          query.$select = request.select.join(",");
-        }
-        const pageResult = await fetchGraphCachePages(graph, {
-          path: request.path,
-          query,
-          headers: request.headers,
-        });
-        store.replaceGraphResources({
-          tenantId: tenant.id,
-          resource,
-          label: definition.label,
-          scopeSet: definition.scopes,
-          rows: pageResult.rows,
-          pageCount: pageResult.pages,
-          pageLimitReached: pageResult.pageLimitReached,
-          refreshedAt,
-        });
-        results.push({
-          resource,
-          label: definition.label,
-          rows: pageResult.rows.length,
-          pages: pageResult.pages,
-          pageLimitReached: pageResult.pageLimitReached,
-          refreshedAt,
-          ok: true,
-        });
-        onProgress?.({
-          type: "resource-complete",
-          result: results.at(-1)!,
-          completed: results.length,
-          total: resources.length,
-        });
-      } catch (error) {
-        const message = error instanceof Error ? error.message : String(error);
-        store.recordGraphResourceError({
-          tenantId: tenant.id,
-          resource,
-          label: definition.label,
-          scopeSet: definition.scopes,
-          error: message,
-          now: refreshedAt,
-        });
-        results.push({
-          resource,
-          label: definition.label,
-          rows: 0,
-          refreshedAt,
-          ok: false,
-          error: message,
-        });
-        onProgress?.({
-          type: "resource-complete",
-          result: results.at(-1)!,
-          completed: results.length,
-          total: resources.length,
-        });
-      }
-    }
-
-    return {
-      tenantId: tenant.id,
-      startedAt,
-      finishedAt: new Date().toISOString(),
-      resources: results,
-    };
+    return this.chatService.refreshGraphCache(options);
   }
 
   async sendIntuneChatMessage(
     input: SendIntuneChatMessageInput,
   ): Promise<SendIntuneChatMessageResult> {
-    const content = input.content.trim();
-    if (!content) {
-      throw new Error("Chat message must not be empty.");
-    }
-
-    const store = this.requireIntelligenceStore();
-    const persisted = await this.read();
-    const tenant = this.resolveTenant(persisted);
-    const providers = await this.listProviders();
-    const provider =
-      providers.find((entry) => entry.id === persisted.activeProviderId) ?? providers[0];
-    const providerId = provider?.id ?? persisted.activeProviderId;
-    const selectedModel = resolveProviderDefaultModel(
-      provider,
-      persisted.activeModelByProviderId,
-    ).model;
-    const chatBudget = intuneChatProviderBudget(providerId);
-    const workspaceContext = input.workspaceContext
-      ? this.buildWorkspacePromptContext(input.workspaceContext, tenant.id, persisted)
-      : undefined;
-    this.requireHostedChatConsent(
-      input,
-      tenant,
-      provider,
-      providerId,
-      workspaceContext?.summary,
-    );
-    const planned = planChatContext(content);
-    const now = new Date().toISOString();
-
-    let conversation = input.conversationId
-      ? store.getConversation(input.conversationId)
-      : undefined;
-    if (conversation) {
-      assertConversationTenant(conversation, tenant.id);
-    }
-    if (!conversation) {
-      conversation = store.createConversation({
-        id: `chat_${randomUUID()}`,
-        title: chatTitleForPrompt(content),
-        tenantId: tenant.id,
-        now,
-      });
-    }
-
-    const userMessage: IntuneChatMessage = {
-      id: `msg_${randomUUID()}`,
-      conversationId: conversation.id,
-      role: "user",
-      content,
-      status: "completed",
-      createdAt: now,
-    };
-    store.insertMessage(userMessage);
-
-    let refreshResult: GraphCacheRefreshResult | undefined;
-    if (input.refreshIfStale !== false) {
-      const before = store.getGraphCacheStatus(tenant.id, [...GRAPH_CACHE_RESOURCES]);
-      const staleResources = planned.resources.filter((resource) => {
-        const status = before.find((entry) => entry.resource === resource);
-        if (!status || status.rows === 0 || !status.refreshedAt) return true;
-        return Date.now() - new Date(status.refreshedAt).getTime() > 6 * 60 * 60 * 1000;
-      });
-      if (staleResources.length > 0) {
-        refreshResult = await this.refreshGraphCache({
-          tenantId: tenant.id,
-          resources: staleResources,
-        });
-        store.insertToolCall({
-          id: `tool_${randomUUID()}`,
-          conversationId: conversation.id,
-          messageId: userMessage.id,
-          type: "graph-cache-refresh",
-          status: refreshResult.resources.some((resource) => resource.ok)
-            ? "completed"
-            : "failed",
-          createdAt: now,
-          completedAt: refreshResult.finishedAt,
-          input: { resources: staleResources },
-          output: refreshResult,
-        });
-      }
-    }
-
-    const cacheStatus = store.getGraphCacheStatus(tenant.id, [...GRAPH_CACHE_RESOURCES]);
-    const answerGeneratedAt = new Date().toISOString();
-    const rows = readPlannedChatRows(store, {
-      tenantId: tenant.id,
-      question: content,
-      generatedAt: answerGeneratedAt,
-      resources: planned.resources,
-      searchTerms: planned.searchTerms,
-      limitPerResource: chatBudget.limitPerResource,
-    });
-    const agentSuggestions = matchAgentsToQuestion(content, persisted.installedAgents);
-    const refreshedResources = new Set(
-      refreshResult?.resources
-        .filter((resource) => resource.ok)
-        .map((resource) => resource.resource) ?? [],
-    );
-    const sources = buildIntuneChatSources({
-      cacheStatus,
-      plannedResources: planned.resources,
-      refreshedResources,
-    });
-
-    const llm = await this.buildLlm(providerId, selectedModel);
-    let assistantContent: string;
-    let assistantStatus: IntuneChatMessage["status"] = "completed";
-    let assistantError: string | undefined;
-
-    if (!llm.available) {
-      assistantStatus = "failed";
-      assistantError =
-        "No LLM provider is connected. Start Ollama or choose another provider in Settings, then try again.";
-      assistantContent = assistantError;
-    } else {
-      const modelQuestion = workspaceContext
-        ? `${content}\n\n${workspaceContext.promptBlock}`
-        : content;
-      const answerPack = buildAnswerPack({
-        question: modelQuestion,
-        tenant,
-        cacheStatus,
-        rows,
-        hasWriteIntent: planned.hasWriteIntent,
-        agentSuggestions,
-        generatedAt: answerGeneratedAt,
-        limits: chatBudget.answerPackLimits,
-      });
-      try {
-        const completion = await llm.complete({
-          system: buildIntuneChatSystemPrompt(provider?.isLocal === true),
-          prompt: `Use this retrieved tenant context to answer the admin.\n\n${answerPack}`,
-          temperature: 0.2,
-          maxTokens: chatBudget.maxTokens,
-        });
-        assistantContent = completion.text.trim();
-        if (planned.hasWriteIntent) {
-          assistantContent = `${assistantContent}\n\nI cannot perform tenant changes directly from chat. Use an installed write agent so the existing plan and confirmation flow stays in force.`;
-        }
-        if (agentSuggestions.length > 0) {
-          assistantContent = `${assistantContent}\n\nDetected matching agent: ${agentSuggestions[0]?.agentName}.`;
-        }
-      } catch (caught) {
-        assistantStatus = "failed";
-        assistantError = caught instanceof Error ? caught.message : String(caught);
-        assistantContent = `The selected LLM provider failed while answering this chat message. ${assistantError}`;
-      }
-    }
-
-    const assistantMessage: IntuneChatMessage = {
-      id: `msg_${randomUUID()}`,
-      conversationId: conversation.id,
-      role: "assistant",
-      content: assistantContent,
-      status: assistantStatus,
-      createdAt: new Date().toISOString(),
-      providerId,
-      model: selectedModel,
-      sources,
-      agentSuggestions,
-      ...(assistantError ? { error: assistantError } : {}),
-    };
-    store.insertMessage(assistantMessage);
-    store.touchConversation(conversation.id, undefined, assistantMessage.createdAt);
-
-    await this.maybeCreateSelfTrainingSuggestionFromChat({
-      tenantId: tenant.id,
-      question: content,
-      agentSuggestions,
-    });
-
-    return {
-      conversation: store.getConversation(conversation.id) ?? conversation,
-      userMessage,
-      assistantMessage,
-      cacheStatus: {
-        tenantId: tenant.id,
-        resources: cacheStatus,
-        schedule: store.getGraphCacheRefreshSchedule(tenant.id),
-      },
-    };
+    return this.chatService.sendIntuneChatMessage(input);
   }
 
   async streamIntuneChatMessage(
@@ -2133,530 +1019,35 @@ export class AppStateStore {
     onEvent: (event: IntuneChatStreamEvent) => void,
     options: { signal?: AbortSignal } = {},
   ): Promise<SendIntuneChatMessageResult> {
-    const content = input.content.trim();
-    if (!content) {
-      throw new Error("Chat message must not be empty.");
-    }
-
-    const store = this.requireIntelligenceStore();
-    const persisted = await this.read();
-    const tenant = this.resolveTenant(persisted);
-    const providers = await this.listProviders();
-    const provider =
-      providers.find((entry) => entry.id === persisted.activeProviderId) ?? providers[0];
-    const providerId = provider?.id ?? persisted.activeProviderId;
-    const selectedModel = resolveProviderDefaultModel(
-      provider,
-      persisted.activeModelByProviderId,
-    ).model;
-    const chatBudget = intuneChatProviderBudget(providerId);
-    const workspaceContext = input.workspaceContext
-      ? this.buildWorkspacePromptContext(input.workspaceContext, tenant.id, persisted)
-      : undefined;
-    this.requireHostedChatConsent(
-      input,
-      tenant,
-      provider,
-      providerId,
-      workspaceContext?.summary,
-    );
-    const planned = planChatContext(content);
-    const now = new Date().toISOString();
-
-    let conversation = input.conversationId
-      ? store.getConversation(input.conversationId)
-      : undefined;
-    if (conversation) {
-      assertConversationTenant(conversation, tenant.id);
-    }
-    if (!conversation) {
-      conversation = store.createConversation({
-        id: `chat_${randomUUID()}`,
-        title: chatTitleForPrompt(content),
-        tenantId: tenant.id,
-        now,
-      });
-    }
-
-    const userMessage: IntuneChatMessage = {
-      id: `msg_${randomUUID()}`,
-      conversationId: conversation.id,
-      role: "user",
-      content,
-      status: "completed",
-      createdAt: now,
-    };
-    store.insertMessage(userMessage);
-
-    const assistantId = `msg_${randomUUID()}`;
-    const pendingAssistant: IntuneChatMessage = {
-      id: assistantId,
-      conversationId: conversation.id,
-      role: "assistant",
-      content: "",
-      status: "streaming",
-      createdAt: new Date().toISOString(),
-      providerId,
-      ...(selectedModel ? { model: selectedModel } : {}),
-    };
-    const initialCacheStatus: GraphCacheStatus = {
-      tenantId: tenant.id,
-      resources: store.getGraphCacheStatus(tenant.id, [...GRAPH_CACHE_RESOURCES]),
-      schedule: store.getGraphCacheRefreshSchedule(tenant.id),
-    };
-    onEvent({
-      type: "started",
-      conversation,
-      userMessage,
-      assistantMessage: pendingAssistant,
-      cacheStatus: initialCacheStatus,
-    });
-    const isCancelled = () => options.signal?.aborted === true;
-    const finishCancelled = (): SendIntuneChatMessageResult => {
-      const cancelledAt = new Date().toISOString();
-      const currentCacheStatus = store.getGraphCacheStatus(tenant.id, [
-        ...GRAPH_CACHE_RESOURCES,
-      ]);
-      const assistantMessage: IntuneChatMessage = {
-        id: assistantId,
-        conversationId: conversation.id,
-        role: "assistant",
-        content: "Response stopped by user before a final answer was saved.",
-        status: "cancelled",
-        createdAt: cancelledAt,
-        providerId,
-        ...(selectedModel ? { model: selectedModel } : {}),
-        sources: buildIntuneChatSources({
-          cacheStatus: currentCacheStatus,
-          plannedResources: planned.resources,
-          refreshedResources: new Set(),
-        }),
-        error: "Stopped by user.",
-      };
-      store.insertMessage(assistantMessage);
-      store.touchConversation(conversation.id, undefined, assistantMessage.createdAt);
-      return {
-        conversation: store.getConversation(conversation.id) ?? conversation,
-        userMessage,
-        assistantMessage,
-        cacheStatus: {
-          tenantId: tenant.id,
-          resources: currentCacheStatus,
-          schedule: store.getGraphCacheRefreshSchedule(tenant.id),
-        },
-      };
-    };
-
-    if (isCancelled()) {
-      const result = finishCancelled();
-      onEvent({ type: "cancelled", result });
-      return result;
-    }
-
-    let refreshResult: GraphCacheRefreshResult | undefined;
-    let progressRefreshResources: GraphCacheResourceKind[] = [];
-    let progressRefreshResults: GraphCacheRefreshResourceResult[] = [];
-    let progressActiveResource: GraphCacheResourceKind | undefined;
-    const sendProgress = (input: {
-      message: string;
-      stage: IntuneChatStreamStage;
-      cacheCheckStatus?: IntuneChatProgressStep["status"];
-      contextStatus?: IntuneChatProgressStep["status"];
-      modelStatus?: IntuneChatProgressStep["status"];
-      cacheStatus?: GraphCacheStatus;
-    }) => {
-      const progressSteps = buildChatProgressSteps({
-        refreshResources: progressRefreshResources,
-        refreshResults: progressRefreshResults,
-        activeResource: progressActiveResource,
-        cacheCheckStatus: input.cacheCheckStatus ?? "completed",
-        contextStatus: input.contextStatus ?? "pending",
-        modelStatus: input.modelStatus ?? "pending",
-      });
-      onEvent({
-        type: "status",
-        conversationId: conversation.id,
-        message: input.message,
-        stage: input.stage,
-        progressSteps,
-        progressPercent: estimateChatProgressPercent(progressSteps),
-        ...(input.cacheStatus ? { cacheStatus: input.cacheStatus } : {}),
-      });
-    };
-    sendProgress({
-      message: "Checking cached tenant data.",
-      stage: "checking-cache",
-      cacheCheckStatus: "active",
-    });
-
-    if (input.refreshIfStale !== false) {
-      const before = store.getGraphCacheStatus(tenant.id, [...GRAPH_CACHE_RESOURCES]);
-      const staleResources = planned.resources.filter((resource) => {
-        const status = before.find((entry) => entry.resource === resource);
-        if (!status || status.rows === 0 || !status.refreshedAt) return true;
-        return Date.now() - new Date(status.refreshedAt).getTime() > 6 * 60 * 60 * 1000;
-      });
-      progressRefreshResources = staleResources;
-      if (staleResources.length > 0) {
-        sendProgress({
-          message: `Refreshing ${staleResources.length} tenant data source${staleResources.length === 1 ? "" : "s"} before answering.`,
-          stage: "refreshing-cache",
-        });
-        refreshResult = await this.refreshGraphCacheInternal(
-          {
-            tenantId: tenant.id,
-            resources: staleResources,
-          },
-          (event) => {
-            if (event.type === "resource-start") {
-              progressActiveResource = event.resource;
-              sendProgress({
-                message: `Refreshing ${event.label}.`,
-                stage: "refreshing-cache",
-              });
-              return;
-            }
-            progressActiveResource = undefined;
-            progressRefreshResults = [...progressRefreshResults, event.result];
-            sendProgress({
-              message: event.result.ok
-                ? `Cached ${event.result.label}.`
-                : `Could not refresh ${event.result.label}.`,
-              stage: "refreshing-cache",
-            });
-          },
-        );
-        store.insertToolCall({
-          id: `tool_${randomUUID()}`,
-          conversationId: conversation.id,
-          messageId: userMessage.id,
-          type: "graph-cache-refresh",
-          status: refreshResult.resources.some((resource) => resource.ok)
-            ? "completed"
-            : "failed",
-          createdAt: now,
-          completedAt: refreshResult.finishedAt,
-          input: { resources: staleResources },
-          output: refreshResult,
-        });
-        sendProgress({
-          message: "Building answer context from local cache.",
-          stage: "building-context",
-          contextStatus: "active",
-          cacheStatus: {
-            tenantId: tenant.id,
-            resources: store.getGraphCacheStatus(tenant.id, [...GRAPH_CACHE_RESOURCES]),
-            schedule: store.getGraphCacheRefreshSchedule(tenant.id),
-          },
-        });
-      } else {
-        sendProgress({
-          message: "Using cached tenant context.",
-          stage: "building-context",
-          contextStatus: "active",
-        });
-      }
-    }
-    if (isCancelled()) {
-      const result = finishCancelled();
-      sendProgress({
-        message: "Response stopped.",
-        stage: "failed",
-        contextStatus: "completed",
-        modelStatus: "failed",
-        cacheStatus: result.cacheStatus,
-      });
-      onEvent({ type: "cancelled", result });
-      return result;
-    }
-
-    const cacheStatus = store.getGraphCacheStatus(tenant.id, [...GRAPH_CACHE_RESOURCES]);
-    const answerGeneratedAt = new Date().toISOString();
-    const rows = readPlannedChatRows(store, {
-      tenantId: tenant.id,
-      question: content,
-      generatedAt: answerGeneratedAt,
-      resources: planned.resources,
-      searchTerms: planned.searchTerms,
-      limitPerResource: chatBudget.limitPerResource,
-    });
-    const agentSuggestions = matchAgentsToQuestion(content, persisted.installedAgents);
-    const refreshedResources = new Set(
-      refreshResult?.resources
-        .filter((resource) => resource.ok)
-        .map((resource) => resource.resource) ?? [],
-    );
-    const sources = buildIntuneChatSources({
-      cacheStatus,
-      plannedResources: planned.resources,
-      refreshedResources,
-    });
-
-    const llm = await this.buildLlm(providerId, selectedModel);
-    let assistantContent = "";
-    let assistantStatus: IntuneChatMessage["status"] = "completed";
-    let assistantError: string | undefined;
-    let responseModel = selectedModel;
-
-    if (!llm.available) {
-      assistantStatus = "failed";
-      assistantError =
-        "No LLM provider is connected. Start Ollama or choose another provider in Settings, then try again.";
-      assistantContent = assistantError;
-    } else {
-      const modelQuestion = workspaceContext
-        ? `${content}\n\n${workspaceContext.promptBlock}`
-        : content;
-      const answerPack = buildAnswerPack({
-        question: modelQuestion,
-        tenant,
-        cacheStatus,
-        rows,
-        hasWriteIntent: planned.hasWriteIntent,
-        agentSuggestions,
-        generatedAt: answerGeneratedAt,
-        limits: chatBudget.answerPackLimits,
-      });
-      try {
-        sendProgress({
-          message: "Model response started.",
-          stage: "generating-answer",
-          contextStatus: "completed",
-          modelStatus: "active",
-          cacheStatus: {
-            tenantId: tenant.id,
-            resources: cacheStatus,
-            schedule: store.getGraphCacheRefreshSchedule(tenant.id),
-          },
-        });
-        for await (const chunk of llm.stream({
-          system: buildIntuneChatSystemPrompt(provider?.isLocal === true),
-          prompt: `Use this retrieved tenant context to answer the admin.\n\n${answerPack}`,
-          ...(selectedModel ? { model: selectedModel } : {}),
-          temperature: 0.2,
-          maxTokens: chatBudget.maxTokens,
-          signal: options.signal,
-        })) {
-          if (isCancelled()) {
-            break;
-          }
-          responseModel = chunk.model;
-          assistantContent = chunk.accumulated;
-          onEvent({
-            type: "delta",
-            conversationId: conversation.id,
-            assistantMessageId: assistantId,
-            delta: chunk.delta,
-            content: assistantContent,
-            providerId,
-            model: responseModel,
-          });
-        }
-        assistantContent = assistantContent.trim();
-        if (planned.hasWriteIntent) {
-          assistantContent = `${assistantContent}\n\nI cannot perform tenant changes directly from chat. Use an installed write agent so the existing plan and confirmation flow stays in force.`;
-        }
-        if (agentSuggestions.length > 0) {
-          assistantContent = `${assistantContent}\n\nDetected matching agent: ${agentSuggestions[0]?.agentName}.`;
-        }
-        onEvent({
-          type: "delta",
-          conversationId: conversation.id,
-          assistantMessageId: assistantId,
-          delta: "",
-          content: assistantContent,
-          providerId,
-          model: responseModel,
-        });
-      } catch (caught) {
-        if (isCancelled()) {
-          const result = finishCancelled();
-          sendProgress({
-            message: "Response stopped.",
-            stage: "failed",
-            contextStatus: "completed",
-            modelStatus: "failed",
-            cacheStatus: result.cacheStatus,
-          });
-          onEvent({ type: "cancelled", result });
-          return result;
-        }
-        assistantStatus = "failed";
-        assistantError = caught instanceof Error ? caught.message : String(caught);
-        assistantContent = assistantContent.trim()
-          ? `${assistantContent.trim()}\n\nThe selected LLM provider failed while answering this chat message. ${assistantError}`
-          : `The selected LLM provider failed while answering this chat message. ${assistantError}`;
-      }
-    }
-    if (isCancelled()) {
-      const result = finishCancelled();
-      sendProgress({
-        message: "Response stopped.",
-        stage: "failed",
-        contextStatus: "completed",
-        modelStatus: "failed",
-        cacheStatus: result.cacheStatus,
-      });
-      onEvent({ type: "cancelled", result });
-      return result;
-    }
-
-    const assistantMessage: IntuneChatMessage = {
-      id: assistantId,
-      conversationId: conversation.id,
-      role: "assistant",
-      content: assistantContent,
-      status: assistantStatus,
-      createdAt: new Date().toISOString(),
-      providerId,
-      ...(responseModel ? { model: responseModel } : {}),
-      sources,
-      agentSuggestions,
-      ...(assistantError ? { error: assistantError } : {}),
-    };
-    store.insertMessage(assistantMessage);
-    store.touchConversation(conversation.id, undefined, assistantMessage.createdAt);
-
-    await this.maybeCreateSelfTrainingSuggestionFromChat({
-      tenantId: tenant.id,
-      question: content,
-      agentSuggestions,
-    });
-
-    const result: SendIntuneChatMessageResult = {
-      conversation: store.getConversation(conversation.id) ?? conversation,
-      userMessage,
-      assistantMessage,
-      cacheStatus: {
-        tenantId: tenant.id,
-        resources: cacheStatus,
-        schedule: store.getGraphCacheRefreshSchedule(tenant.id),
-      },
-    };
-
-    sendProgress({
-      message:
-        assistantStatus === "failed"
-          ? "Chat answer failed."
-          : "Chat answer ready.",
-      stage: assistantStatus === "failed" ? "failed" : "completed",
-      contextStatus: "completed",
-      modelStatus: assistantStatus === "failed" ? "failed" : "completed",
-      cacheStatus: result.cacheStatus,
-    });
-
-    if (assistantStatus === "failed") {
-      onEvent({
-        type: "failed",
-        result,
-        error: assistantError ?? "The selected LLM provider failed.",
-      });
-    } else {
-      onEvent({ type: "completed", result });
-    }
-    return result;
-  }
-
-  private requireHostedChatConsent(
-    input: SendIntuneChatMessageInput,
-    tenant: TenantRecord,
-    provider: ProviderSummary | undefined,
-    providerId: ProviderId,
-    workspaceContext?: WorkspacePromptContextSummary,
-  ): void {
-    if (provider?.isLocal !== false) {
-      return;
-    }
-
-    const consent = input.hostedProviderConsent;
-    if (!consent) {
-      throw new Error(
-        "Hosted provider confirmation is required before Intune Chat can send tenant context to the selected provider.",
-      );
-    }
-
-    if (consent.tenantId !== tenant.id || consent.providerId !== providerId) {
-      throw new Error(
-        "Hosted provider confirmation does not match the active tenant and provider. Confirm the chat send again.",
-      );
-    }
-
-    const acknowledgedAt = Date.parse(consent.acknowledgedAt);
-    const now = Date.now();
-    if (
-      !Number.isFinite(acknowledgedAt) ||
-      now - acknowledgedAt > 5 * 60 * 1000 ||
-      acknowledgedAt - now > 60 * 1000
-    ) {
-      throw new Error(
-        "Hosted provider confirmation expired. Confirm the chat send again.",
-      );
-    }
-
-    if (!workspaceContext) {
-      return;
-    }
-
-    const acknowledgedWorkspace = consent.workspaceContext;
-    if (
-      !acknowledgedWorkspace ||
-      acknowledgedWorkspace.workspaceId !== workspaceContext.workspaceId ||
-      acknowledgedWorkspace.tenantId !== workspaceContext.tenantId ||
-      acknowledgedWorkspace.evidenceCount !== workspaceContext.evidenceCount ||
-      acknowledgedWorkspace.noteCount !== workspaceContext.noteCount ||
-      acknowledgedWorkspace.includesInstructions !== workspaceContext.includesInstructions
-    ) {
-      throw new Error(
-        "Hosted provider confirmation does not include the attached workspace context. Confirm the chat send again.",
-      );
-    }
+    return this.chatService.streamIntuneChatMessage(input, onEvent, options);
   }
 
   async getSelfTrainingSettings(): Promise<SelfTrainingSettings> {
-    return this.requireIntelligenceStore().getSelfTrainingSettings();
+    return this.chatService.getSelfTrainingSettings();
   }
 
   async setSelfTrainingEnabled(enabled: boolean): Promise<SelfTrainingSettings> {
-    const now = new Date().toISOString();
-    return this.requireIntelligenceStore().setSelfTrainingEnabled(enabled, now);
+    return this.chatService.setSelfTrainingEnabled(enabled);
   }
 
   async listSelfTrainingSuggestions(
     status?: SelfTrainingSuggestionStatus,
   ): Promise<SelfTrainingSuggestion[]> {
-    return this.requireIntelligenceStore().listSelfTrainingSuggestions(status);
+    return this.chatService.listSelfTrainingSuggestions(status);
   }
 
   async approveSelfTrainingSuggestion(id: string): Promise<SelfTrainingSuggestion> {
-    const suggestion = this.requireIntelligenceStore().decideSelfTrainingSuggestion(
-      id,
-      "accepted",
-      new Date().toISOString(),
-    );
-    await this.writeSelfTrainingFile(suggestion.tenantId, suggestion.agentSlug);
-    return suggestion;
+    return this.chatService.approveSelfTrainingSuggestion(id);
   }
 
   async rejectSelfTrainingSuggestion(id: string): Promise<SelfTrainingSuggestion> {
-    return this.requireIntelligenceStore().decideSelfTrainingSuggestion(
-      id,
-      "rejected",
-      new Date().toISOString(),
-    );
+    return this.chatService.rejectSelfTrainingSuggestion(id);
   }
 
   async resetSelfTrainingSuggestions(
     input: ResetSelfTrainingInput,
   ): Promise<SelfTrainingSuggestion[]> {
-    const persisted = await this.read();
-    const tenantId = this.resolveTenant(persisted, input.tenantId).id;
-    const reset = this.requireIntelligenceStore().resetAcceptedSelfTrainingSuggestions({
-      tenantId,
-      agentSlug: input.agentSlug,
-      now: new Date().toISOString(),
-    });
-    await this.writeSelfTrainingFile(tenantId, input.agentSlug);
-    return reset;
+    return this.chatService.resetSelfTrainingSuggestions(input);
   }
 
   async listWorkspaces(tenantId?: string): Promise<WorkspaceSummary[]> {
@@ -3644,41 +2035,7 @@ export class AppStateStore {
   }
 
   async refreshDueGraphCaches(): Promise<void> {
-    if (!this.intelligenceStore) return;
-    const persisted = await this.read();
-    const nowMs = Date.now();
-
-    for (const tenant of persisted.tenants) {
-      if (!this.intelligenceStore.isGraphCacheRefreshDue(tenant.id, nowMs)) continue;
-      const startedAt = new Date(nowMs).toISOString();
-      try {
-        const result = await this.refreshGraphCache({ tenantId: tenant.id });
-        const failures = result.resources.filter((resource) => !resource.ok);
-        this.intelligenceStore.markGraphCacheRefreshScheduleRun({
-          tenantId: tenant.id,
-          startedAt,
-          success: failures.length === 0,
-          ...(failures.length > 0
-            ? {
-                error: failures
-                  .map((failure) => `${failure.label}: ${failure.error ?? "failed"}`)
-                  .join("; "),
-              }
-            : {}),
-        });
-      } catch (error) {
-        this.intelligenceStore.markGraphCacheRefreshScheduleRun({
-          tenantId: tenant.id,
-          startedAt,
-          success: false,
-          error: error instanceof Error ? error.message : String(error),
-        });
-        console.error(
-          `[intune-chat] scheduled Graph cache refresh failed for tenant ${tenant.id}:`,
-          error,
-        );
-      }
-    }
+    return this.chatService.refreshDueGraphCaches();
   }
 
   private async persistScheduledRunFailure(
